Lots of modifications
Made the program loop faster by removing the pressure and humidity
compensated calculations until the time of upload.  This has
significantly improved the performance of the program (roughly doubled).

Replaced a lot of the constants which were used for calculations with
magic numbers so that the readability is improved.

Renamed methods and variables for easier readability.

Next step is to check the code still works with the sensors and then
merge with the WeatherStation sketch.